+++ core/index/index.go
@@ -11,11 +11,13 @@ import (
 type ElasticsearchConfig struct {
 	Endpoint    string `config:"endpoint"`
 	IndexPrefix string `config:"index_prefix"`
+	Username    string `config:"username"`
+	Password    string `config:"password"`
 }
 
 // ElasticsearchClient elasticsearch client api
 type ElasticsearchClient struct {
-	Config ElasticsearchConfig
+	Config *ElasticsearchConfig
 }
 
 // InsertResponse is a index response object
@@ -81,12 +83,14 @@ func (query *RangeQuery) Gte(field string, value interface{}) {
 	v["gte"] = value
 	query.Range[field] = v
 }
+
 func (query *RangeQuery) Lt(field string, value interface{}) {
 	query.Range = map[string]map[string]interface{}{}
 	v := map[string]interface{}{}
 	v["lt"] = value
 	query.Range[field] = v
 }
+
 func (query *RangeQuery) Lte(field string, value interface{}) {
 	query.Range = map[string]map[string]interface{}{}
 	v := map[string]interface{}{}
@@ -151,15 +155,17 @@ func (c *ElasticsearchClient) Index(indexName, id string, data interface{}) (*In
 	if err != nil {
 		return nil, err
 	}
-	response := util.HttpPostJSON(url, "", string(js))
+	req := util.NewPostRequest(url, js)
+	req.SetBasicAuth(c.Config.Username, c.Config.Password)
+	response, err := util.ExecuteRequest(req)
 	if err != nil {
 		return nil, err
 	}
 
-	log.Trace("indexing response: ", string(response))
+	log.Trace("indexing response: ", string(response.Body))
 
 	esResp := &InsertResponse{}
-	err = json.Unmarshal(response, esResp)
+	err = json.Unmarshal(response.Body, esResp)
 	if err != nil {
 		return &InsertResponse{}, err
 	}
@@ -176,15 +182,18 @@ func (c *ElasticsearchClient) Get(indexName, id string) (*GetResponse, error) {
 
 	log.Debug("get doc: ", url)
 
-	response, err := util.HttpGet(url)
+	req := util.NewGetRequest(url)
+	req.SetBasicAuth(c.Config.Username, c.Config.Password)
+	response, err := util.ExecuteRequest(req)
+
 	if err != nil {
 		return nil, err
 	}
 
-	log.Trace("get response: ", string(response))
+	log.Trace("get response: ", string(response.Body))
 
 	esResp := &GetResponse{}
-	err = json.Unmarshal(response, esResp)
+	err = json.Unmarshal(response.Body, esResp)
 	if err != nil {
 		return &GetResponse{}, err
 	}
@@ -201,15 +210,17 @@ func (c *ElasticsearchClient) Delete(indexName, id string) (*DeleteResponse, err
 
 	log.Debug("delete doc: ", url)
 
-	response, err := util.HttpDelete(url)
+	req := util.NewDeleteRequest(url)
+	req.SetBasicAuth(c.Config.Username, c.Config.Password)
+	response, err := util.ExecuteRequest(req)
 	if err != nil {
 		return nil, err
 	}
 
-	log.Trace("delete response: ", string(response))
+	log.Trace("delete response: ", string(response.Body))
 
 	esResp := &DeleteResponse{}
-	err = json.Unmarshal(response, esResp)
+	err = json.Unmarshal(response.Body, esResp)
 	if err != nil {
 		return &DeleteResponse{}, err
 	}
@@ -228,15 +239,18 @@ func (c *ElasticsearchClient) Count(indexName string) (*CountResponse, error) {
 
 	log.Debug("doc count: ", url)
 
-	response, err := util.HttpGet(url)
+	req := util.NewGetRequest(url)
+	req.SetBasicAuth(c.Config.Username, c.Config.Password)
+	response, err := util.ExecuteRequest(req)
+
 	if err != nil {
 		return nil, err
 	}
 
-	log.Trace("count response: ", string(response))
+	log.Trace("count response: ", string(response.Body))
 
 	esResp := &CountResponse{}
-	err = json.Unmarshal(response, esResp)
+	err = json.Unmarshal(response.Body, esResp)
 	if err != nil {
 		return &CountResponse{}, err
 	}
@@ -263,13 +277,21 @@ func (c *ElasticsearchClient) Search(indexName string, query *SearchRequest) (*S
 	}
 
 	js, err := json.Marshal(query)
+	if err != nil {
+		return nil, err
+	}
 
-	response := util.HttpPostJSON(url, "", string(js))
+	req := util.NewPostRequest(url, js)
+	req.SetBasicAuth(c.Config.Username, c.Config.Password)
+	response, err := util.ExecuteRequest(req)
+	if err != nil {
+		return nil, err
+	}
 
-	log.Trace("search response: ", string(js), ",", string(response))
+	log.Trace("search response: ", string(js), ",", string(response.Body))
 
 	esResp := &SearchResponse{}
-	err = json.Unmarshal(response, esResp)
+	err = json.Unmarshal(response.Body, esResp)
 	if err != nil {
 		return &SearchResponse{}, err
 	}