+++ core/model/task.go
@@ -16,26 +16,32 @@ import (
 type TaskStatus int
 
 const TaskCreated TaskStatus = 0
-const TaskFetchFailed TaskStatus = 2
-const TaskFetchSuccess TaskStatus = 3
-const Task404Ignore TaskStatus = 4
-const TaskRedirectedIgnore TaskStatus = 5
-const TaskFetchTimeout TaskStatus = 6
+const TaskFailed TaskStatus = 2
+const TaskSuccess TaskStatus = 3
+const Task404 TaskStatus = 4
+const TaskRedirected TaskStatus = 5
+const TaskTimeout TaskStatus = 6
+const TaskDuplicated TaskStatus = 7
+const TaskInterrupted TaskStatus = 8
 
 func GetTaskStatusText(status TaskStatus) string {
 	switch status {
 	case TaskCreated:
 		return "created"
-	case TaskFetchFailed:
+	case TaskFailed:
 		return "failed"
-	case Task404Ignore:
+	case Task404:
 		return "404"
-	case TaskFetchSuccess:
+	case TaskSuccess:
 		return "success"
-	case TaskRedirectedIgnore:
+	case TaskRedirected:
 		return "redirected"
-	case TaskFetchTimeout:
+	case TaskTimeout:
 		return "timeout"
+	case TaskDuplicated:
+		return "duplicated"
+	case TaskInterrupted:
+		return "interrupted"
 	}
 	return "unknow"
 }
@@ -122,18 +128,18 @@ type Task struct {
 	OriginalUrl string          `json:"original_url,omitempty"`
 	Phrase      pipeline.Phrase `gorm:"index" json:"phrase"`
 	Status      TaskStatus      `gorm:"index" json:"status"`
-	Message     string          `json:"-"`
+	Message     string          `json:"message,omitempty"`
 	Created     *time.Time      `gorm:"index" json:"created,omitempty"`
 	Updated     *time.Time      `gorm:"index" json:"updated,omitempty"`
-	LastFetch   *time.Time      `gorm:"index" json:"last_fetch"`
-	LastCheck   *time.Time      `gorm:"index" json:"last_check"`
-	NextCheck   *time.Time      `gorm:"index" json:"next_check"`
+	LastFetch   *time.Time      `gorm:"index" json:"last_fetch,omitempty"`
+	LastCheck   *time.Time      `gorm:"index" json:"last_check,omitempty"`
+	NextCheck   *time.Time      `gorm:"index" json:"next_check,omitempty"`
 
-	SnapshotVersion int        `json:"snapshot_version"`
-	SnapshotID      string     `json:"snapshot_id"`
-	SnapshotHash    string     `json:"snapshot_hash"`
-	SnapshotSimHash string     `json:"snapshot_simhash"`
-	SnapshotCreated *time.Time `json:"snapshot_created"`
+	SnapshotVersion int        `json:"snapshot_version,omitempty"`
+	SnapshotID      string     `json:"snapshot_id,omitempty"`
+	SnapshotHash    string     `json:"snapshot_hash,omitempty"`
+	SnapshotSimHash string     `json:"snapshot_simhash,omitempty"`
+	SnapshotCreated *time.Time `json:"snapshot_created,omitempty"`
 }
 
 func CreateTask(task *Task) error {
@@ -251,7 +257,7 @@ func GetPendingUpdateFetchTasks(offset *time.Time) (int, []Task, error) {
 		Conds: persist.And(
 			persist.Lt("next_check", t),
 			persist.Gt("created", offset),
-			persist.Eq("status", TaskFetchSuccess)),
+			persist.Eq("status", TaskSuccess)),
 		From: 0, Size: 100,
 	}
 	err, result := persist.Search(Task{}, &tasks, &queryO)