+++ core/util/idutil.go
@@ -27,10 +27,13 @@ type incrementCounter struct {
 var count = incrementCounter{l: &sync.RWMutex{}, ID: make(map[string]*atomicID)}
 
 type atomicID struct {
+	l        sync.Mutex
 	Sequence int64
 }
 
 func (id *atomicID) Increment() int64 {
+	id.l.Lock()
+	defer id.l.Unlock()
 	return atomic.AddInt64(&id.Sequence, 1)
 }
 
@@ -53,8 +56,8 @@ func GetIncrementID(bucket string) string {
 
 // SnapshotPersistID will make a snapshot and persist id stats to disk
 func SnapshotPersistID() {
-	lock1.Lock()
-	defer lock1.Unlock()
+	count.l.Lock()
+	defer count.l.Unlock()
 
 	var buf bytes.Buffer
 	err := gob.NewEncoder(&buf).Encode(count)