+++ core/util/webhunter_test.go
@@ -17,18 +17,18 @@ limitations under the License.
 package util
 
 import (
-	"encoding/json"
 	"fmt"
 	"github.com/magiconair/properties/assert"
 	"testing"
 )
 
 func TestGet(t *testing.T) {
-	data, _ := get("http://es-guide-preview.elasticsearch.cn", "", "")
-
-	data1, _ := json.Marshal(data)
-	fmt.Println("", string(data1))
-	//assert.Equal(t,data.StatusCode,301)
+	data, err := get("http://es-guide-preview.elasticsearch.cn", "", "")
+	fmt.Println(data)
+	fmt.Println(err)
+	//data1, _ := json.Marshal(data.Body)
+	//fmt.Println("", string(data1))
+	//assert.Equal(t, data.StatusCode, 301)
 }
 
 func TestGetHost(t *testing.T) {
@@ -72,4 +72,12 @@ func TestGetHost(t *testing.T) {
 	host = GetHost(url)
 	fmt.Println("logo.com", host)
 	assert.Equal(t, host, "logo.com")
+}
+
+func BenchmarkGet(b *testing.B) {
+
+	for i := 0; i < b.N; i++ {
+		get("http://es-guide-preview.elasticsearch.cn", "", "")
+	}
+
 }