+++ modules/api/handlers/handler.go
@@ -37,8 +37,9 @@ func InitAPI() {
 	api.HandleAPIMethod(api.GET, "/", apis.IndexAction)
 	api.HandleAPIMethod(api.GET, "/favicon.ico", apis.IndexAction)
 
-	//APIs
+	//Stats APIs
 	api.HandleAPIFunc("/stats", apis.StatsAction)
+	api.HandleAPIFunc("/queue/stats", apis.QueueStatsAction)
 
 	//Task API
 	api.HandleAPIMethod(api.GET, "/tasks/", apis.TaskAction)