+++ modules/crawler/pipe/parse_test.go
@@ -18,14 +18,19 @@ package pipe
 
 import (
 	"fmt"
+	"github.com/infinitbyte/gopa/core/env"
+	"github.com/infinitbyte/gopa/core/global"
 	"github.com/infinitbyte/gopa/core/model"
 	"github.com/infinitbyte/gopa/core/pipeline"
 	"github.com/infinitbyte/gopa/core/util"
 	"github.com/stretchr/testify/assert"
+	"io/ioutil"
 	"testing"
 )
 
 func TestProcessLinks(t *testing.T) {
+	global.RegisterEnv(env.EmptyEnv())
+
 	body := "<!DOCTYPE html> <html> <head> <meta content=\"text/html;charset=utf-8\" http-equiv=\"Content-Type\" /> <meta content=\"width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no\" name=\"viewport\" /> <meta http-equiv=\"X-UA-Compatible\" content=\"IE=edge,Chrome=1\" /> <meta name=\"renderer\" content=\"webkit\" /> <title>Elastic中文社区</title> <meta name=\"keywords\" content=\"Elasticsearch中文社区，实时数据分析，实时数据检索, Elastic Stack，ELK，elasticsearch、logstash、kibana、beats等相关技术交流探讨\" /> <meta name=\"description\" content=\"Elasticsearch中文社区，elasticsearch、logstash、kibana,beats等相关技术交流探讨\" /> <base href=\"http://elasticsearch.cn/\" /><!--[if IE]></base><![endif]--> <link href=\"http://elasticsearch.cn/static/css/default/img/favicon.ico?v=20151125\" rel=\"shortcut icon\" type=\"image/x-icon\" /> <link rel=\"stylesheet\" type=\"text/css\" href=\"http://elasticsearch.cn/static/css/bootstrap.css\" /> <link rel=\"stylesheet\" type=\"text/css\" href=\"http://elasticsearch.cn/static/css/icon.css\" /> <link href=\"http://elasticsearch.cn/static/css/default/common.css?v=20151125\" rel=\"stylesheet\" type=\"text/css\" /> <link href=\"http://elasticsearch.cn/static/css/default/link.css?v=20151125\" rel=\"stylesheet\" type=\"text/css\" /> <link href=\"http://elasticsearch.cn/static/js/plug_module/style.css?v=20151125\" rel=\"stylesheet\" type=\"text/css\" /> </head> <body> <div style=\"display:none;\" id=\"__crond\"><a href=\"google.com\">myLink</a>" +
 		"<a href=\"//baidu.com\">baidu</a>" +
 		"<h1>h1<span>123</span></h1>" +
@@ -48,7 +53,10 @@ func TestProcessLinks(t *testing.T) {
 
 	context.Set(CONTEXT_CRAWLER_TASK, &task)
 	parse := ParsePageJoint{}
+	parse.Set(dispatchLinks, false)
 	snapshot := model.Snapshot{}
+	snapshot.ContentType = "text/html"
+
 	context.Set(CONTEXT_CRAWLER_SNAPSHOT, &snapshot)
 	snapshot.Payload = []byte(body)
 	parse.Process(&context)
@@ -60,4 +68,59 @@ func TestProcessLinks(t *testing.T) {
 	assert.Equal(t, "/wiki/Marking/Users", links["/wiki/Marking/Users"])
 	assert.Equal(t, "myLink", links["google.com"])
 	fmt.Println(util.ToJson(snapshot, true))
+
+	//load file
+	b, e := ioutil.ReadFile("../../../test/samples/discuss.html")
+	if e != nil {
+		panic(e)
+	}
+	task = model.Task{}
+	task.Url = "http://discuss.elastic.co/"
+	task.Host = "discuss.elastic.co"
+
+	context.Set(CONTEXT_CRAWLER_TASK, &task)
+	snapshot = model.Snapshot{}
+	snapshot.ContentType = "text/html"
+
+	context.Set(CONTEXT_CRAWLER_SNAPSHOT, &snapshot)
+	snapshot.Payload = []byte(b)
+	parse.Process(&context)
+
+	o = context.MustGet(CONTEXT_PAGE_LINKS)
+	fmt.Println("links in discuss:")
+	fmt.Println(util.ToJson(o, true))
+
+}
+func TestProcessDiscussLinks(t *testing.T) {
+	global.RegisterEnv(env.EmptyEnv())
+
+	body, e := ioutil.ReadFile("../../../test/samples/discuss.html")
+	if e != nil {
+		panic(e)
+	}
+
+	context := pipeline.Context{}
+	context.Init()
+	task := model.Task{}
+	task.Url = "http://discuss.elastic.co/"
+	task.Host = "discuss.elastic.co"
+
+	context.Set(CONTEXT_CRAWLER_TASK, &task)
+	parse := ParsePageJoint{}
+	parse.Set(dispatchLinks, false)
+	snapshot := model.Snapshot{}
+	snapshot.ContentType = "text/html"
+
+	context.Set(CONTEXT_CRAWLER_SNAPSHOT, &snapshot)
+	snapshot.Payload = []byte(body)
+	parse.Process(&context)
+
+	o := context.MustGet(CONTEXT_PAGE_LINKS)
+	links := o.(map[string]string)
+	//println(links["google.com"])
+	//assert.Equal(t, "baidu", links["//baidu.com"])
+	//assert.Equal(t, "/wiki/Marking/Users", links["/wiki/Marking/Users"])
+	//assert.Equal(t, "myLink", links["google.com"])
+	fmt.Println(util.ToJson(links, true))
+
 }