+++ modules/crawler/pipeline.go
@@ -49,6 +49,9 @@ func InitJoints() {
 	pipeline.Register(pipe.HashJoint{})
 	pipeline.Register(pipe.IndexJoint{})
 	pipeline.Register(pipe.TaskDeduplicationJoint{})
+	pipeline.Register(pipe.ContentDeduplicationJoint{})
+	pipeline.Register(pipe.UpdateCheckTimeJoint{})
+	pipeline.Register(pipe.LanguageDetectJoint{})
 	log.Debug("end register joints")
 
 }