+++ modules/persist/elastic/elasticsearch.go
@@ -14,6 +14,7 @@ type ElasticORM struct {
 func getIndex(any interface{}) string {
 	return util.GetTypeName(any, true)
 }
+
 func getID(any interface{}) string {
 	return util.GetFieldValueByTagName(any, "index", "id")
 }